# lib/api-zod/src/generated/types/dealer.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Dealer {
  id: number;
  name: string;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  xmlFeedUrl?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  lastSyncAt?: string | null;
  /** @nullable */
  lastSyncStatus?: string | null;
  totalVehiclesImported: number;
  /** @nullable */
  lastError?: string | null;
  createdAt: string;
}